wafr-agents/wafr/agents/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
anthropic>=0.18.0
beautifulsoup4>=4.12.0
requests>=2.31.0
requests-html>=0.10.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
openai-whisper>=20230314

